dashboard/requirements.txt
streamlit>=1.30.0
psycopg2-binary>=2.9.9
pandas>=1.5.3
